Create guidebook-worksheets.md
As discussed in today's meeting #16, I have started this document with the goal of mapping the current guidebook sections to the new class outline. I realize that we only discussed mapping titles to worksheets/video/remove, but I also added a few notes about in-class demos / pre-reading materials (it felt intuitive to do this as I was reading through the guidebook); I hope this is okay! My impression is that this could be a useful outline going forward, I also thought that it may be a good place to include a "learning objectives" section within each revised class meeting?

I have tried to keep it organized, but I realize that there are some ideas that would probably better fit elsewhere (i.e. the addition of the syllabus as a part of the course structure section within the website). I will make note of these and add them into the general 2020 new outline document as well. I look forward to hearing your thoughts, and if you find that this type of document works well with our goal, and seems helpful for thinking about the classes / videos / worksheets, I am happy to continue slowly chipping away at it! Otherwise we can revisit and discuss what's best.

Thank you,
Icíar