Create mini-project.md
I have started this document to gain clarity on the mini data analysis project, as discussed in #16. So far, it is quite preliminary and I have tried to stick to a structured document that we can modify as we discuss - once the learning objectives and overarching framework of the project are defined, the natural first step might be to start creating the deliverables. Parallel to this, it may be a good idea to simultaneously work on hashing out the details of the collaborative troubleshooting assignments to schedule deliverables in the least overwhelming way possible to students - however, I do think this project should take more time because it's worth more % in the course. 

In addition to what's detailed in the document, another alternative could also be making "deliverable 4" be a final project report where students can summarize their workflow and findings in an R Markdown document, and instead summarize deliverables 1 and 2 into a single deliverable (if becoming familiarized with dplyr seems too easy for a full deliverable, which it may be). I look forward to hearing what you think!

Thank you,
Icíar