Create 2020-resources.md
As discussed in #16, I have started this ongoing doc for resources, the idea being that it will eventually live under the "Resources" tab in the STAT545 website. So far, I have migrated all resources found in the current STAT545 guidebook and divided them into sections; I am now working on structuring each section as I've done with the table in the Data Visualization section, under "R". I have also added a few resources that I have found useful in learning R, and the idea with this page is to modify it as more resources are shared in different years of STAT545. 

I was also wondering if it would be worth it to add a "Stats" section at the very end for those students that may want to explore that further, or if that's beyond the scope of STAT545. I look forward to hearing your thoughts on the document so far.

Thank you,
Icíar