Add files via upload
Here is a draft of the first worksheet! I have yet to add the section that will correspond to Class 3: R Markdown & Reproducibility; so far I have only designed the first part corresponding to Class 1. I would like to acknowledge that although part of it is completely created from scratch, I have also taken some sections from the current guidebook & adapted them, as well as used the swirl() R Programming and the "A very (short) introduction to R" document by Claudia Brauer as inspiration (not copy-paste but did make some similar exercises adapting them to the workflow of the worksheet!). I hope this is okay, and I would also like to note it somewhere in the worksheet (maybe a "References" section at the end?) because I think it's important to credit others' work - but I wanted to consult you on how is best to do this. I look forward to hearing your thoughts!

Going forward, I think that it may be easier if I learn how to knit the files to Jupyter notebooks and implement nbGrader, I tried to make notes throughout for Team Sonja where the nbGrader sections would be, so hopefully that's enough but please feel free to reach out if anything needs clarification (and same goes for anyone working in Team Sonja). Thank you!