reorganize team troubleshooting documents
Once PR #36 is closed and merged, this PR should be merged next so that work on these documents (if required) can continue in their new location - for some reason I had two folders "materials-2020" and "2020-materials" in my local copy of this repo, at some point I must've fetched a branch too early, and created an additional folder by mistake. "materials-2020" already exists in the Master branch of this repo, and so I will continue keeping documents in that folder, and delete "2020-materials" altogether (which is the folder that merging #36 will create).