mini project deliverable 1 + instructions
Hi all,

I have written a draft for the first deliverable of the Mini Data Analysis project, and added overall instructions for the whole project. As discussed in #21, the hope with the project is for students to practice on their own independent data analysis while being provided some sort of framework / guidance to minimize open-endness and define the scope of the project (also to avoid some students spending too much time on it, and others too little). @Vincenzo suggested the following overall "topics" for the deliverables:

1. Rmd and some dplyr exploration.
2. Tables (through tidy data manipulation)
3. Graphs (perhaps plotting those tables they made in Milestone 2, and this would also be useful for making the point that it's almost always better to present a graph than a table.)

I'd appreciate any feedback on this deliverable, and please don't hesitate to ask me any questions! Oh and where I say in Task 3 that they can see an example, I'm working on that still - just a separate brief document that they can have a look at.